+++ src/__tests__/App.test.jsx
@@ -1,11 +1,20 @@
 import { describe, expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
-import { MemoryRouter } from "react-router-dom";
+import { createBrowserRouter } from "react-router-dom";
 import App from "../App";
+import routes from "../routes/routes";
 
 describe("App component", () => {
-  it("The App component is rendered", () => {
-    const { container } = render(<App />);
+  it("The App component matches snapshot", () => {
+    const router = createBrowserRouter(routes, { initialEntries: ["/"] });
+    const { container } = render(<App router={router} />);
     expect(container).toMatchSnapshot();
+  });
+
+  it("Renders splash screen on load and while API request is in progress", () => {
+    const router = createBrowserRouter(routes, { initialEntries: ["/"] });
+    render(<App router={router} />);
+    const loading = screen.getByTitle("Loading");
+    expect(loading).toBeInTheDocument();
   });
 });