+++ src/__tests__/Categories.test.jsx
@@ -1,4 +1,4 @@
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
 import Categories from "../components/Categories";
@@ -13,17 +13,6 @@ describe("Categories component", () => {
     );
 
     expect(container).toMatchSnapshot();
-  });
-
-  it("The Categories component without categories prop renders Loading component", () => {
-    render(
-      <MemoryRouter initialEntries={["/shop"]}>
-        <Categories />
-      </MemoryRouter>
-    );
-
-    const categories = screen.getByTitle("Loading");
-    expect(categories).toBeInTheDocument();
   });
 
   it("The categories list should have 5 items", () => {