+++ src/__tests__/ErrorPage.test.jsx
@@ -1,19 +1,35 @@
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, vi } from "vitest";
 import { render, screen } from "@testing-library/react";
+import { RouterProvider, createMemoryRouter } from "react-router-dom";
 import ErrorPage from "../components/ErrorPage";
+import routes from "../routes/routes";
 
 describe("ErrorPage component", () => {
   it("The ErrorPage component is rendered", () => {
-    render(<ErrorPage />);
-    // Why can't I use getByRole('region')?
-    // const errorPageSection = screen.getByText("Error");
+    const router = createMemoryRouter(routes, { initialEntries: ["/foo"] });
+
+    render(
+      <RouterProvider router={router}>
+        <ErrorPage />
+      </RouterProvider>
+    );
+
     const errorPageSection = screen.getByRole("region");
     expect(errorPageSection).toBeInTheDocument();
   });
 
-  it('The ErrorPage component text is "404 Page not found"', () => {
-    const errorMessage = "404 Page not found";
-    render(<ErrorPage errorMessage={errorMessage} />);
-    expect(screen.getByText(errorMessage)).toHaveTextContent(errorMessage);
+  it('The ErrorPage component text is "404 Page not found"', async () => {
+    const errorMessage = "404 - Page not found";
+    const fetchSpy = vi.spyOn(globalThis, "fetch");
+    fetchSpy.mockResolvedValue({
+      status: 404,
+      message: errorMessage,
+    });
+
+    const router = createMemoryRouter(routes, { initialEntries: ["/"] });
+
+    render(<RouterProvider router={router} />);
+    const errorPageSection = await screen.findByRole("region");
+    expect(errorPageSection).toHaveTextContent(errorMessage);
   });
 });