+++ src/__tests__/Footer.test.jsx
@@ -1,11 +1,10 @@
 import { describe, expect, it } from "vitest";
-import { render, screen } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import Footer from "../components/Footer";
 
 describe("Footer component", () => {
-  it("The Footer component is rendered", () => {
-    render(<Footer />);
-    const footer = screen.getByRole("contentinfo");
-    expect(footer).toBeInTheDocument();
+  it("The Footer component matches snapshot", () => {
+    const { container } = render(<Footer />);
+    expect(container).toMatchSnapshot();
   });
 });