+++ src/__tests__/Shop.test.jsx
@@ -1,13 +1,9 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import {
-  render,
-  screen,
-  waitForElementToBeRemoved,
-} from "@testing-library/react";
-import { MemoryRouter, Route, Routes } from "react-router-dom";
+import { render, screen } from "@testing-library/react";
+import { RouterProvider, createMemoryRouter } from "react-router-dom";
 import { fetchMock } from "./mocks";
-import Shop from "../components/Shop";
-import parseProducts from "../utilities/parseProducts";
+import { CartProvider } from "../components/Cart";
+import routes from "../routes/routes";
 
 /* Optional
  * Define fetchMock here
@@ -26,61 +22,28 @@ beforeEach(() => {
 });
 
 describe("Shop component", () => {
-  it("The mock function, fetchMock, is called twice", () => {
-    const spyFetch = vi.spyOn(globalThis, "fetch");
-    render(
-      <MemoryRouter initialEntries={["/shop"]}>
-        <Shop />
-      </MemoryRouter>
-    );
-
-    expect(spyFetch).toBeCalledTimes(2);
-  });
-
-  it("The mock function, parseProductsMock, is called once", () => {
-    render(
-      <MemoryRouter initialEntries={["/shop"]}>
-        <Routes>
-          <Route path="/shop" element={<Shop />} />
-        </Routes>
-      </MemoryRouter>
-    );
-
-    expect(parseProducts).toBeCalledTimes(1);
-  });
-
   it("The Shop component is rendered", async () => {
+    const router = createMemoryRouter(routes, { initialEntries: ["/shop"] });
     render(
-      <MemoryRouter initialEntries={["/shop"]}>
-        <Routes>
-          <Route path="/shop" element={<Shop />} />
-        </Routes>
-      </MemoryRouter>
+      <CartProvider>
+        <RouterProvider router={router} />
+      </CartProvider>
     );
     const shopSection = await screen.findByText("Shop section");
     expect(shopSection).toBeInTheDocument();
   });
 
-  it("Renders correct text while API request is in progress", async () => {
-    render(
-      <MemoryRouter initialEntries={["/"]}>
-        <Shop />
-      </MemoryRouter>
-    );
-    const categories = screen.getByTitle("Loading");
-    expect(categories).toBeInTheDocument();
-    await waitForElementToBeRemoved(() => screen.getByTitle("Loading"));
-  });
-
   it("Renders correct text when fetch fails", async () => {
-    window.fetch.mockImplementationOnce(() =>
-      Promise.reject({ message: "API is down" })
+    const rejectCause = { status: 404 };
+    window.fetch.mockRejectedValue(
+      new Error("API is down", { cause: { ...rejectCause } })
     );
+    const router = createMemoryRouter(routes, { initialEntries: ["/shop"] });
 
     render(
-      <MemoryRouter initialEntries={["/"]}>
-        <Shop />
-      </MemoryRouter>
+      <CartProvider>
+        <RouterProvider router={router} />
+      </CartProvider>
     );
 
     const errorMessage = await screen.findByText("API is down");
@@ -88,14 +51,18 @@ describe("Shop component", () => {
   });
 
   it("The 'All' category is first in the categories list", async () => {
+    const router = createMemoryRouter(routes, { initialEntries: ["/shop"] });
     render(
-      <MemoryRouter initialEntries={["/"]}>
-        <Shop />
-      </MemoryRouter>
+      <CartProvider>
+        <RouterProvider router={router} />
+      </CartProvider>
     );
 
-    const categoriesList = await screen.findByRole("list");
-    const categoriesListItems = categoriesList.children;
+    const lists = await screen.findAllByRole("list");
+    const categoriesListItems = lists.find((list) =>
+      list.textContent.match(/all/i)
+    ).children;
+
     const firstCategory = categoriesListItems[0];
     expect(firstCategory.textContent).match(/all/i);
   });