+++ src/components/Loading.jsx
@@ -1,9 +1,13 @@
 import LoadingIcon from "../assets/icons/progress_activity.svg?react";
+import styles from "../styles/Loading.module.css";
 
 const Loading = () => {
   return (
-    <div>
-      <LoadingIcon />
+    <div className={styles["loading"]}>
+      <div className={styles["loading-wrapper"]}>
+        <LoadingIcon className={styles["loading-icon"]} />
+        {/* <p>Application loading</p> */}
+      </div>
     </div>
   );
 };