+++ src/routes/loader.js
@@ -1,14 +1,28 @@
-import { defer } from "react-router-dom";
 import parseProducts from "../utilities/parseProducts";
 
 const loader = async () => {
   const data = await Promise.all([
     fetch("https://fakestoreapi.com/products/categories", { mode: "cors" }),
     fetch("https://fakestoreapi.com/products", { mode: "cors" }),
+    // Promise.reject(new Error("error")),
+    // Promise.reject(new Error("error")),
+    // Promise.resolve({
+    //   status: 400,
+    //   message: "Something broke",
+    // }),
+    // Promise.resolve({
+    //   status: 500,
+    // }),
+    // fetch(""),
+    // fetch(""),
   ])
-    .then(([resCategories, resProducts]) => {
+    .then((responses) => {
       // How to mock response status codes >= 400?
-      if (resCategories.status >= 400) throw new Error("ERROR");
+      // How to refactor the if blocks for status codes?
+      const invalidResponse = responses.find((res) => res.status >= 400);
+      // console.log(invalidResponse);
+      invalidResponse && throwException(invalidResponse);
+      const [resCategories, resProducts] = responses;
       return Promise.all([resCategories.json(), resProducts.json()]);
     })
     .then(([dataCategories, dataProducts]) => {
@@ -18,29 +32,20 @@ const loader = async () => {
         products: parsedProducts,
       };
     })
-    .catch((error) => error);
+    .catch((error) => {
+      throw {
+        ...error,
+        error: new Error(error.message || error),
+        // ...(error.status && { status: error.status }),
+      };
+    });
+  // .catch((error) => ({ error: { cause: error } }));
+
   return data;
 };
 
-// const loader = async () => {
-//   const data = Promise.all([
-//     fetch("https://fakestoreapi.com/products/categories", { mode: "cors" }),
-//     fetch("https://fakestoreapi.com/products", { mode: "cors" }),
-//   ])
-//     .then(([resCategories, resProducts]) => {
-//       // How to mock response status codes >= 400?
-//       if (resCategories.status >= 400) throw new Error("ERROR");
-//       return Promise.all([resCategories.json(), resProducts.json()]);
-//     })
-//     .then(([dataCategories, dataProducts]) => {
-//       const parsedProducts = parseProducts(dataProducts);
-//       return {
-//         categories: ["all", ...dataCategories],
-//         products: parsedProducts,
-//       };
-//     })
-//     .catch((error) => error);
-//   return defer({ data });
-// };
+const throwException = (invalidResponse) => {
+  throw invalidResponse;
+};
 
 export default loader;